--- ee/tabby-ui/app/(dashboard)/(logs)/jobs/utils/state.ts
@@ -1,5 +1,0 @@
-import { isNil } from 'lodash-es'
-
-export function getLabelByExitCode(exitCode?: number | null) {
-  return isNil(exitCode) ? 'Pending' : exitCode === 0 ? 'Success' : 'Failed'
-}
